# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: BangMusic Server API
  version: 1.0.0
  summary: 자가호스팅 BangMusic 음악 서버와 모바일 앱 사이의 계약
  description: |
    설계 초안(2026-10-02). 구현 전 계약이며 동작하는 서버가 아직 없다.

    공통 규칙
    - 모든 경로는 `/v1` 아래에 있다. 호환되지 않는 변경은 `/v2`로만 낸다.
    - `/v1` 안에서는 추가만 한다(새 필드·새 엔드포인트·새 enum 값). 앱은 모르는 필드와 enum 값을 무시해야 한다.
    - 서버가 무엇을 지원하는지는 `GET /v1/server`의 `api.minor`와 `features`로 판단한다.
    - 오류는 RFC 9457 `application/problem+json`이며 기계 판독용 `code`를 포함한다.
    - 응답에는 서버의 파일 경로가 절대 포함되지 않는다. 모든 자원은 불투명 ID로만 가리킨다.
    - 시간은 RFC 3339 UTC 문자열, 길이는 밀리초 정수다.
    - 권한이 없는 자원은 존재 여부를 숨기기 위해 404 `not_found`로 응답한다. 403은 역할이 부족한 관리 API에만 쓴다.
    - 모든 응답에 `X-Request-Id`가 있다. 앱은 오류 보고에 이 값을 표시한다.
    - 모든 동작은 `426 client_too_old`(앱 API minor가 서버 최소값보다 낮음)와 `503`(`setup_required`, `maintenance` 등)을 돌려줄 수 있다.
  license:
    name: 미정 (서버 패키지 라이선스는 제품 소유자 결정 사항)
    identifier: LicenseRef-TBD
servers:
  - url: https://{host}/v1
    description: 구매자가 운영하는 서버. 앱은 사용자가 입력한 주소를 사용한다.
    variables:
      host:
        default: music.example.net
...
security:
  - bearerAuth: []
paths:
  /server:
    get:
      tags: [server]
      operationId: getServerInfo
      summary: 서버 식별자·API 버전·기능 목록
      description: |
        앱이 서버 주소를 등록할 때와 앱 시작 시마다 호출한다. 인증 없이 호출 가능하며
        라이브러리 내용이나 사용자 정보는 포함하지 않는다.
      security: []
      responses:
        '200':
          description: 서버 정보
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerInfo' }
              example:
                server_id: srv_01JB2M4X7Q9ZK3V5T8N6R1WY0E
                name: 우리집 음악
                product: bangmusic-server
                version: 1.0.0
                api: { major: 1, minor: 0 }
                min_client_api_minor: 0
                features: [transcode, lyrics, playlist_covers, history, change_feed]
                transcode_profiles: [aac_256, aac_128]
                limits: { max_page_size: 200, max_history_batch: 500, max_cover_bytes: 5242880 }
                setup_required: false
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
...
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: 리프레시 토큰을 회전시키고 새 토큰 쌍을 발급
      description: |
        리프레시 토큰은 1회용이다. 직전 토큰은 네트워크 실패 후 재시도를 위해 60초 동안만 다시 받아 주며,
        이때 같은 새 토큰 쌍을 돌려준다. 그 이후에 옛 토큰이 다시 쓰이면 탈취로 보고 세션 전체를 폐기한다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: 새 토큰 쌍
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: |
            `refresh_expired`(만료), `session_revoked`(관리자·사용자가 취소 또는 재사용 감지),
            `account_disabled`. 재시도해도 성공하지 않으므로 앱은 재로그인을 요구한다.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
...
  /sessions:
    get:
      tags: [auth]
      operationId: listSessions
      summary: 내 기기 세션 목록
      responses:
        '200':
          description: 세션 목록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /sessions/{sessionId}:
    delete:
      tags: [auth]
      operationId: revokeSession
      summary: 내 다른 기기의 세션 취소
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: 취소됨 (멱등) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /me/password:
    put:
      tags: [me]
      operationId: changePassword
      summary: 비밀번호 변경 (다른 모든 세션 폐기)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 10, maxLength: 256 }
      responses:
        '204': { description: 변경됨 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /tracks:
    get:
      tags: [catalog]
      operationId: listTracks
      summary: 곡 목록
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: sort, in: query, schema: { type: string, enum: [title, added_at_desc, album], default: title } }
        - { name: album_id, in: query, schema: { type: string } }
        - { name: artist_id, in: query, schema: { type: string } }
        - { name: library_id, in: query, schema: { type: string } }
      responses:
        '200':
          description: 곡 페이지
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /tracks/{trackId}:
    get:
      tags: [catalog]
      operationId: getTrack
      summary: 곡 상세 조회
      parameters:
        - { $ref: '#/components/parameters/TrackId' }
      responses:
        '200':
          description: 곡
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Track' }
              example:
                id: trk_01JB2P3C5E7G9J1L3N5Q7S9U1W
                title: 夜明けのうた
                title_sort: よあけのうた
                artists: [{ id: art_01JB2P1A3C5E7G9J1L3N5Q7S9U, name: テスト楽団 }]
                album: { id: alb_01JB2P2B4D6F8H0K2M4P6R8T0V, title: 合成アルバム }
                disc_no: 1
                track_no: 3
                duration_ms: 215340
                year: 2024
                artwork_id: img_01JB2P0Z2B4D6F8H0K2M4P6R8T
                media_version: k7q2m9x4b1c8d5f0
                source_format: { container: flac, codec: flac, sample_rate: 44100, channels: 2, bit_depth: 16, bitrate: 912000 }
                state: available
                has_lyrics: [original, pronunciation_ko, translation_ko]
                updated_at: '2026-09-28T11:02:13Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /albums:
    get:
      tags: [catalog]
      operationId: listAlbums
      summary: 앨범 목록 (커서 페이지네이션)
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: sort, in: query, schema: { type: string, enum: [title, added_at_desc, year_desc], default: title } }
        - { name: artist_id, in: query, schema: { type: string } }
      responses:
        '200':
          description: 앨범 페이지
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /albums/{albumId}:
    get:
      tags: [catalog]
      operationId: getAlbum
      summary: 앨범과 수록곡 (디스크·트랙 순)
      parameters:
        - { name: albumId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 앨범 상세
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /artists:
    get:
      tags: [catalog]
      operationId: listArtists
      summary: 아티스트 목록 (커서 페이지네이션)
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: 아티스트 페이지
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArtistPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /artists/{artistId}:
    get:
      tags: [catalog]
      operationId: getArtist
      summary: 아티스트 상세와 앨범 목록
      parameters:
        - { name: artistId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 아티스트
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artist' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /changes:
    get:
      tags: [catalog]
      operationId: getChanges
      summary: 증분 동기화용 변경 피드
      parameters:
        - name: since
          in: query
          required: true
          description: 이전 응답의 `next_since`. 최초 동기화는 `0`.
          schema: { type: string }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: 변경 묶음
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangeBatch' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '410':
          description: '`sync_token_expired` — `since`가 보존 기간보다 오래됨. 앱은 전체 재동기화한다.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /artwork/{artworkId}:
    get:
      tags: [artwork]
      operationId: getArtwork
      summary: 표지 이미지
      description: |
        `artwork_id`는 내용이 바뀌면 새 ID가 되므로 응답은 불변으로 캐시할 수 있다.
        `Authorization` 헤더 또는 `mt` 미디어 티켓으로 인증한다.
      security:
        - bearerAuth: []
        - mediaTicket: []
      parameters:
        - { name: artworkId, in: path, required: true, schema: { type: string } }
        - name: size
          in: query
          description: 긴 변의 픽셀 수. 서버가 가장 가까운 준비된 크기로 내린다.
          schema: { type: integer, enum: [96, 256, 512, 1024], default: 512 }
      responses:
        '200':
          description: 이미지
          headers:
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string, examples: ['private, max-age=31536000, immutable'] } }
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
        '304': { description: 변경 없음 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /tracks/{trackId}/renditions:
    post:
      tags: [media]
      operationId: resolveRendition
      summary: 이 기기에서 재생/다운로드할 제공 형식 결정
      description: |
        앱이 재생 가능한 형식과 원하는 품질을 보내면 서버가 원본 제공 또는 변환을 결정한다.
        - 원본을 그대로 줄 수 있거나 변환본이 캐시에 있으면 200과 `state=ready`.
        - 변환이 필요하면 작업을 대기열에 넣고 202와 `state=preparing`, `Retry-After`.
          앱은 `GET /renditions/{id}`를 폴링한다.
        같은 요청을 반복해도 같은 렌디션이 나온다(멱등). 응답의 `media_url`은 서버 기준 상대 경로이며
        `mt` 미디어 티켓이 포함되어 있어 플레이어·다운로더가 헤더 없이 요청할 수 있다.
      parameters:
        - { $ref: '#/components/parameters/TrackId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenditionRequest' }
            example:
              purpose: stream
              quality: original
              accept: [{ container: flac, codec: flac }, { container: mp3, codec: mp3 }, { container: m4a, codec: aac }, { container: ogg, codec: opus }, { container: ogg, codec: vorbis }]
      responses:
        '200':
          description: 바로 사용 가능
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rendition' }
              example:
                id: rnd_01JB2R7G9J1L3N5Q7S9U1W3Y5A
                track_id: trk_01JB2P3C5E7G9J1L3N5Q7S9U1W
                media_version: k7q2m9x4b1c8d5f0
                profile: original
                state: ready
                mime: audio/flac
                container: flac
                codec: flac
                size_bytes: 24561892
                duration_ms: 215340
                sha256: 9f2c1e0d6b7a48c3a5d4e1f09b8c7d6e5f4a3b2c1d0e9f8a7b6c5d4e3f2a1b0c
                etag: '"k7q2m9x4b1c8d5f0-original"'
                seekable: true
                media_url: /v1/media/rnd_01JB2R7G9J1L3N5Q7S9U1W3Y5A?mt=eyJ2IjoxLC...redacted
                ticket_expires_at: '2026-10-02T15:00:00Z'
        '202':
          description: 변환 준비 중
          headers:
            Retry-After: { schema: { type: integer }, description: 다음 폴링까지 권장 초 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rendition' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: '`not_found`(곡 없음/권한 없음) 또는 `media_missing`(곡은 있으나 원본 파일이 사라짐)'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: '`no_playable_format` — 앱이 받을 수 있는 형식이 없고 서버 변환도 꺼져 있음'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429':
          description: '`transcode_queue_full` — 변환 대기열 한도 초과. `Retry-After` 후 재시도.'
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
  /renditions/{renditionId}:
    get:
      tags: [media]
      operationId: getRendition
      summary: 렌디션 상태 조회 및 미디어 티켓 재발급
      description: |
        `preparing` 상태 폴링과, 티켓이 만료된 뒤 새 `media_url`을 받는 데 쓴다.
        원본이 바뀌어 이 렌디션이 더 이상 유효하지 않으면 410 `rendition_superseded`.
      parameters:
        - { $ref: '#/components/parameters/RenditionId' }
      responses:
        '200':
          description: 렌디션 (`state`는 `ready`, `preparing`, `failed` 중 하나)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rendition' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/Gone' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /media/{renditionId}:
    parameters:
      - { $ref: '#/components/parameters/RenditionId' }
    get:
      tags: [media]
      operationId: getMedia
      summary: 오디오 바이트 (스트리밍과 다운로드 공용)
      description: |
        렌디션은 불변이다. 같은 ID는 항상 같은 바이트를 가리키며, 원본이 바뀌면 410이 된다.
        따라서 이어받기에서 옛 바이트와 새 바이트가 섞일 수 없다.
        단일 범위의 `Range`만 지원한다. 여러 범위를 요청하면 전체(200)를 돌려준다.
      security:
        - bearerAuth: []
        - mediaTicket: []
      parameters:
        - name: Range
          in: header
          schema: { type: string, examples: ['bytes=1048576-'] }
        - name: If-Range
          in: header
          description: 렌디션의 ETag. 일치하지 않으면 200 전체 응답.
          schema: { type: string }
      responses:
        '200':
          description: 전체 본문
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer } }
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string, examples: ['private, no-transform, max-age=31536000, immutable'] } }
            X-Content-Duration-Ms: { schema: { type: integer } }
          content:
            audio/*: { schema: { type: string, format: binary } }
        '206':
          description: 부분 본문
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Range: { schema: { type: string, examples: ['bytes 1048576-24561891/24561892'] } }
            Content-Length: { schema: { type: integer } }
            ETag: { schema: { type: string } }
          content:
            audio/*: { schema: { type: string, format: binary } }
        '401':
          description: '`ticket_expired`, `ticket_invalid`, `access_expired`, `session_revoked` 중 하나'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`rendition_not_ready` — 아직 변환 중'
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '410': { $ref: '#/components/responses/Gone' }
        '416':
          description: '`range_not_satisfiable`'
          headers:
            Content-Range: { schema: { type: string, examples: ['bytes */24561892'] } }
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
    head:
      tags: [media]
      operationId: headMedia
      summary: 길이·MIME·ETag 확인 (본문 없음)
      security:
        - bearerAuth: []
        - mediaTicket: []
      responses:
        '200':
          description: GET과 같은 헤더
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer } }
            Content-Type: { schema: { type: string } }
            ETag: { schema: { type: string } }
            X-Content-Duration-Ms: { schema: { type: integer } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/Gone' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /admin/users:
    get:
      tags: [admin]
      operationId: adminListUsers
      summary: 사용자 목록 (관리자)
      responses:
        '200':
          description: 사용자 목록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AdminUser' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [admin]
      operationId: adminCreateUser
      summary: 사용자 생성 (관리자)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string, minLength: 1, maxLength: 64 }
                display_name: { type: string, maxLength: 100 }
                password: { type: string, minLength: 10, maxLength: 256 }
                role: { type: string, enum: [admin, member] }
                library_ids: { type: array, items: { type: string } }
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /admin/users/{userId}:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string } }
    patch:
      tags: [admin]
      operationId: adminUpdateUser
      summary: 역할·상태·라이브러리 접근·비밀번호 재설정
      description: '`status=disabled` 또는 라이브러리 접근 축소는 즉시 적용되며 해당 사용자의 미디어 티켓도 무효가 된다.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                role: { type: string, enum: [admin, member] }
                status: { type: string, enum: [active, disabled] }
                library_ids: { type: array, items: { type: string } }
                new_password: { type: string, minLength: 10, maxLength: 256 }
                revoke_sessions: { type: boolean }
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      tags: [admin]
      operationId: adminDeleteUser
      summary: 사용자 삭제 (관리자)
      responses:
        '204': { description: 삭제됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /admin/libraries:
    get:
      tags: [admin]
      operationId: adminListLibraries
      summary: 라이브러리와 스캔 상태 (경로는 포함하지 않음)
      responses:
        '200':
          description: 라이브러리 목록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AdminLibrary' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /admin/libraries/{libraryId}/scan:
    post:
      tags: [admin]
      operationId: adminStartScan
      summary: 스캔 시작
      parameters:
        - { name: libraryId, in: path, required: true, schema: { type: string } }
      responses:
        '202':
          description: 작업 접수. 이미 실행 중이면 그 작업을 돌려준다.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
  /admin/jobs/{jobId}:
    get:
      tags: [admin]
      operationId: adminGetJob
      summary: 관리 작업 상태 조회 (관리자)
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 작업 상태
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <access_token>`. 액세스 토큰은 불투명 문자열이며 서버가 해시로 조회한다.'
    mediaTicket:
      type: apiKey
      in: query
      name: mt
      description: |
        렌디션·표지 요청 전용 단기 서명 티켓. 사용자·세션·대상 자원·만료 시각에 묶인다.
        다른 API에는 쓸 수 없다. 서버와 프록시는 접근 로그에서 `mt` 값을 가린다.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: 이전 응답의 `next_cursor`. 불투명 문자열이며 앱이 만들거나 해석하지 않는다.
      schema: { type: string }
...
  responses:
...
    RateLimited:
      description: '`rate_limited` — `Retry-After` 초 뒤 재시도'
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unavailable:
      description: |
        `storage_unavailable`(음악 저장소 마운트 장애), `maintenance`(마이그레이션·복원 중),
        `setup_required`(관리자 계정 미생성). 지수 백오프로 재시도 가능.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...